perf: zero alloc hashing (#21)
As seen in the flamegraph, dozens of **billions** of objects are being heap allocated while hashing. This PR aims to bring it down to zero.
<img width="1234" height="469" alt="image" src="https://github.com/user-attachments/assets/b6a42b34-12ab-4390-8bbe-49ca7706cb78" />

The culprit is `hash.Sum`, which allocates a new slice if `nil` is given to it. The solution is to provide a real slice to it, so it copies instead of allocating. Also, the tree now allocates a single contiguous set of arrays, colocated in memory, rather than using slices. Altogether, ***this brings allocations ~95% down***.
                                                      
```
                                  |  allocs/op    |     allocs/op     |   delta
  NewTree/leaves_64-16            136.0 ± 0%           9.000 ± 0%       -93.4%  
  NewTree/leaves_256-16           558.0 ± 0%           47.00 ± 0%       -91.6%  
  NewTree/leaves_1024-16         2.134k ± 0%           85.00 ± 0%       -96.0%  
  NewTree/leaves_4096-16         8.322k ± 0%           124.0 ± 0%       -98.5%  
  geomean                        1.077k                45.95            -95.7%
```

The profiles on high load are to follow.